backend/Recommend/requirements.txt
numpy<2.0.0
pandas
scikit-learn
scipy
surprise
Flask
pika
python-dotenv